# doc/contracts/backend-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: LabRat Backend API
  version: 1.0.0
  description: >-
    Contract-first API for the NestJS modular monolith. All documented
    operations have closed DTOs and authoritative v1 behavior.
    Public Guest identities are restricted to one demo project with read/export
    access. Authenticated Guest writes, AI requests and invitation redemption
    return 403 public_guest_read_only; only GET/HEAD and POST logout are allowed.
    Guest request limits may return 429 too_many_requests on authenticated reads.
servers:
  - url: http://127.0.0.1:8788
    description: Local v1 development server
...
security:
  - cookieAuth: []
paths:
  /api/v1/admin/invitations:
    get:
      tags: [Identity]
      operationId: listOwnerInvitations
      parameters:
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationPage" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "410": { description: "Invitation expired, revoked or used.", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "413": { description: Request body exceeds 8 KiB., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "429": { description: Too many attempts., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    post:
      tags: [Identity]
      operationId: createOwnerInvitation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmptyInvitationRequest" }
      responses:
        "201":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedInvitation" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "410": { description: "Invitation expired, revoked or used.", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "413": { description: Request body exceeds 8 KiB., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "429": { description: Too many attempts., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /api/v1/admin/invitations/{invitationId}/revoke:
    parameters:
      - { name: invitationId, in: path, required: true, schema: { $ref: "#/components/schemas/OpaqueId" } }
    post:
      tags: [Identity]
      operationId: revokeOwnerInvitation
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "410": { description: "Invitation expired, revoked or used.", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "413": { description: Request body exceeds 8 KiB., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "429": { description: Too many attempts., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /api/v1/labs/{labId}/invitations:
    parameters: [{ $ref: "#/components/parameters/LabId" }]
    get:
      tags: [Identity]
      operationId: listMemberInvitations
      parameters:
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationPage" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "410": { description: "Invitation expired, revoked or used.", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "413": { description: Request body exceeds 8 KiB., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "429": { description: Too many attempts., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    post:
      tags: [Identity]
      operationId: createMemberInvitation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmptyInvitationRequest" }
      responses:
        "201":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedInvitation" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "410": { description: "Invitation expired, revoked or used.", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "413": { description: Request body exceeds 8 KiB., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "429": { description: Too many attempts., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /api/v1/labs/{labId}/invitations/{invitationId}/revoke:
    parameters:
      - { $ref: "#/components/parameters/LabId" }
      - { name: invitationId, in: path, required: true, schema: { $ref: "#/components/schemas/OpaqueId" } }
    post:
      tags: [Identity]
      operationId: revokeMemberInvitation
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "410": { description: "Invitation expired, revoked or used.", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "413": { description: Request body exceeds 8 KiB., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "429": { description: Too many attempts., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /api/v1/auth/invitations/preview:
    post:
      tags: [Identity]
      operationId: previewInvitation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationCodeRequest" }
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationPreview" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "410": { description: "Invitation expired, revoked or used.", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "413": { description: Request body exceeds 8 KiB., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "429": { description: Too many attempts., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /api/v1/auth/register:
    post:
      tags: [Identity]
      operationId: registerWithInvitation
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterInvitationRequest" }
      responses:
        "201":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationRedemption" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "410": { description: "Invitation expired, revoked or used.", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "413": { description: Request body exceeds 8 KiB., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "429": { description: Too many attempts., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /api/v1/auth/invitations/redeem:
    post:
      tags: [Identity]
      operationId: redeemInvitation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RedeemInvitationRequest" }
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationRedemption" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "410": { description: "Invitation expired, revoked or used.", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "413": { description: Request body exceeds 8 KiB., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "429": { description: Too many attempts., content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: labrat_session
  parameters:
...
    LabId:
      name: labId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/OpaqueId" }
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, maxLength: 500 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
  responses:
    Unauthorized:
      description: Authentication is missing or expired.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: The actor lacks the required capability.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Conflict:
      description: Version, state, stale-head or idempotency conflict.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationError:
      description: Request failed bounded validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    EmptyInvitationRequest:
      type: object
      additionalProperties: false
      properties: {}
    InvitationCodeRequest:
      type: object
      additionalProperties: false
      required: [invitationCode]
      properties:
        invitationCode: { type: string, pattern: "^[A-Za-z0-9_-]{43}$", writeOnly: true }
    RedeemInvitationRequest:
      type: object
      additionalProperties: false
      required: [invitationCode]
      properties:
        invitationCode: { type: string, pattern: "^[A-Za-z0-9_-]{43}$", writeOnly: true }
        labName: { type: string, minLength: 1, maxLength: 300 }
    RegisterInvitationRequest:
      type: object
      additionalProperties: false
      required: [invitationCode, username, displayName, password]
      properties:
        invitationCode: { type: string, pattern: "^[A-Za-z0-9_-]{43}$", writeOnly: true }
        labName: { type: string, minLength: 1, maxLength: 300 }
        username: { type: string, minLength: 1, maxLength: 200 }
        displayName: { type: string, minLength: 1, maxLength: 300 }
        password: { type: string, minLength: 12, maxLength: 1000, writeOnly: true }
    Invitation:
      type: object
      additionalProperties: false
      required: [id, kind, labId, createdBy, createdAt, expiresAt, status, revokedAt, redeemedAt, redeemedBy, redeemedLabId]
      properties:
        id: { $ref: "#/components/schemas/OpaqueId" }
        kind: { type: string, enum: [lab_owner, lab_member] }
        labId: { $ref: "#/components/schemas/NullableOpaqueId" }
        createdBy: { $ref: "#/components/schemas/OpaqueId" }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        expiresAt: { $ref: "#/components/schemas/Timestamp" }
        status: { type: string, enum: [pending, used, revoked, expired] }
        revokedAt: { type: [string, "null"], format: date-time }
        redeemedAt: { type: [string, "null"], format: date-time }
        redeemedBy: { $ref: "#/components/schemas/NullableOpaqueId" }
        redeemedLabId: { $ref: "#/components/schemas/NullableOpaqueId" }
    InvitationPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Invitation" } }
        nextCursor: { $ref: "#/components/schemas/NullableString" }
    CreatedInvitation:
      type: object
      additionalProperties: false
      required: [invitation, invitationCode]
      properties:
        invitation: { $ref: "#/components/schemas/Invitation" }
        invitationCode: { type: string, description: Shown once. Never persist or put in URLs. }
    InvitationResponse:
      type: object
      additionalProperties: false
      required: [invitation]
      properties:
        invitation: { $ref: "#/components/schemas/Invitation" }
    InvitationPreview:
      type: object
      additionalProperties: false
      required: [kind, expiresAt, lab]
      properties:
        kind: { type: string, enum: [lab_owner, lab_member] }
        expiresAt: { $ref: "#/components/schemas/Timestamp" }
        lab:
          oneOf:
            - { type: "null" }
            - type: object
              additionalProperties: false
              required: [id, name]
              properties:
                id: { $ref: "#/components/schemas/OpaqueId" }
                name: { type: string }
    InvitationRedemption:
      type: object
      additionalProperties: false
      required: [auth, lab]
      properties:
        auth: { $ref: "#/components/schemas/AuthResponse" }
        lab: { $ref: "#/components/schemas/Lab" }
...
    OpaqueId:
      type: string
      minLength: 1
      maxLength: 200
    Timestamp:
      type: string
      format: date-time
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: { type: string, minLength: 1, maxLength: 100 }
            message: { type: string, minLength: 1, maxLength: 1000 }
            requestId: { type: string, minLength: 1, maxLength: 200 }
            details:
              type: object
              additionalProperties: true
...
    User:
      type: object
      additionalProperties: false
      required: [id, username, displayName, isActive, isSuperAdmin]
      properties:
        id: { $ref: "#/components/schemas/OpaqueId" }
        username: { type: string }
        displayName: { type: string }
        isActive: { type: boolean }
        isSuperAdmin: { type: boolean }
    LabMembership:
      type: object
      additionalProperties: false
      required: [labId, role, status]
      properties:
        labId: { $ref: "#/components/schemas/OpaqueId" }
        role: { type: string, enum: [lab_owner, lab_admin, lab_member] }
        status: { type: string, enum: [active, inactive] }
...
    AuthResponse:
      type: object
      additionalProperties: false
      required: [user, memberships]
      properties:
        user: { $ref: "#/components/schemas/User" }
        memberships:
          type: array
          items: { $ref: "#/components/schemas/LabMembership" }
...
    Lab:
      type: object
      additionalProperties: false
      required: [id, name, slug, status]
      properties:
        id: { $ref: "#/components/schemas/OpaqueId" }
        name: { type: string }
        slug: { type: string }
        status: { type: string, enum: [active, inactive] }
        role: { type: string, enum: [lab_owner, lab_admin, lab_member] }
...
    NullableString:
      oneOf:
        - type: string
        - type: "null"
...
    NullableOpaqueId:
      oneOf:
        - { $ref: "#/components/schemas/OpaqueId" }
        - type: "null"
...